feat(fish): install pure prompt, add macOS platform guards
- Install fisher + pure prompt via fisher
- Guard Arch/Linux-only aliases (pacman, grub, journalctl, etc.)
- Guard rofi/scrot/zapret aliases in config.fish
- Fall back to osascript for topdf notifications on macOS
- Guard dmenu-scripts path additions

+++ fish/.config/fish/cachyos-config.fish
@@ -83,8 +83,6 @@ alias lt='eza -aT --color=always --group-directories-first --icons' # tree listi
 alias l.="eza -a | grep -e '^\.'" # show only dotfiles
 
 # Common use
-alias grubup="sudo grub-mkconfig -o /boot/grub/grub.cfg"
-alias fixpacman="sudo rm /var/lib/pacman/db.lck"
 alias tarnow='tar -acf '
 alias untar='tar -zxvf '
 alias wget='wget -c '
@@ -95,30 +93,26 @@ alias ...='cd ../..'
 alias ....='cd ../../..'
 alias .....='cd ../../../..'
 alias ......='cd ../../../../..'
-alias dir='dir --color=auto'
-alias vdir='vdir --color=auto'
 alias grep='grep --color=auto'
 alias fgrep='fgrep --color=auto'
 alias egrep='egrep --color=auto'
-alias hw='hwinfo --short' # Hardware Info
-alias big="expac -H M '%m\t%n' | sort -h | nl" # Sort installed packages according to size in MB
-alias gitpkg='pacman -Q | grep -i "\-git" | wc -l' # List amount of -git packages
-alias update='sudo pacman -Syu'
-
-# Get fastest mirrors
-alias mirror="sudo cachyos-rate-mirrors"
-
-# Help people new to Arch
-alias apt='man pacman'
-alias apt-get='man pacman'
 alias please='sudo'
 alias tb='nc termbin.com 9999'
 
-# Cleanup orphaned packages
-alias cleanup='sudo pacman -Rns (pacman -Qtdq)'
-
-# Get the error messages from journalctl
-alias jctl="journalctl -p 3 -xb"
-
-# Recent installed packages
-alias rip="expac --timefmt='%Y-%m-%d %T' '%l\t%n %v' | sort | tail -200 | nl"
+# Linux / Arch-only aliases
+if test (uname) = Linux
+    alias grubup="sudo grub-mkconfig -o /boot/grub/grub.cfg"
+    alias fixpacman="sudo rm /var/lib/pacman/db.lck"
+    alias dir='dir --color=auto'
+    alias vdir='vdir --color=auto'
+    alias hw='hwinfo --short' # Hardware Info
+    alias big="expac -H M '%m\t%n' | sort -h | nl" # Sort installed packages according to size in MB
+    alias gitpkg='pacman -Q | grep -i "\-git" | wc -l' # List amount of -git packages
+    alias update='sudo pacman -Syu'
+    alias mirror="sudo cachyos-rate-mirrors" # Get fastest mirrors
+    alias apt='man pacman' # Help people new to Arch
+    alias apt-get='man pacman'
+    alias cleanup='sudo pacman -Rns (pacman -Qtdq)' # Cleanup orphaned packages
+    alias jctl="journalctl -p 3 -xb" # Get the error messages from journalctl
+    alias rip="expac --timefmt='%Y-%m-%d %T' '%l\t%n %v' | sort | tail -200 | nl" # Recent installed packages
+end
